iOS: extract CoverThumbnailView, unify presets, and clean up readability
Batch 1 - Cleanups:
- Remove unused import Combine from EmulatorBridge.swift
- Remove stale cleanup-history comments from ios_main.mm
- Simplify StorageCleaner.report(paths:) to delegate directly to
  the private awaitableReport(paths:) overload, eliminating the
  duplicate synchronous wrapper
- Route scattered StikDebug direct print calls through a new private
  StikDebugLauncher.log(_:) helper in EmulatorBridge.swift

Batch 2 - Component Extractions:
- Extract CoverThumbnailView and CoverThumbnailCache out of GameListView
  into a dedicated Views/CoverThumbnailView.swift; register the new
  source file in CMakeLists.txt
- Unify duplicate compatibility preset tables and metadata into a single
  shared CompatibilityPreset Swift model; consumed by both GameScreenView
  and GameListView, removing the duplicated literal tables

Also fix Metal shader include paths for fxaa and cas: replace broken
PC-relative paths (../../../../bin/resources/...) with
MTL_HEADER_SEARCH_PATHS-relative includes (shaders/common/...).
