+++ app/src/main/cpp/pcsx2/GS/Renderers/Metal/fxaa.metal
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0+
 
 #include "GSMTLShaderCommon.h"
-#include "../../../../bin/resources/shaders/common/fxaa.fx"
+#include "shaders/common/fxaa.fx"
 
 fragment float4 ps_fxaa(ConvertShaderData data [[stage_in]], texture2d<float> tex [[texture(GSMTLTextureIndexNonHW)]])
 {
